add cluster inplace upgrade e2e test
