+++ test/e2e/config/docker.yaml
@@ -86,8 +86,8 @@ variables:
   # This avoids building node images in the default case which improves the test duration significantly.
   KUBERNETES_VERSION_MANAGEMENT: "v1.29.0"
   KUBERNETES_VERSION: "v1.29.0"
-  KUBERNETES_VERSION_UPGRADE_FROM: "v1.28.0"
-  KUBERNETES_VERSION_UPGRADE_TO: "v1.29.0"
+  KUBERNETES_VERSION_UPGRADE_FROM: "v1.29.0"
+  KUBERNETES_VERSION_UPGRADE_TO: "v1.29.1"
   ETCD_VERSION_UPGRADE_TO: "3.5.10-0"
   COREDNS_VERSION_UPGRADE_TO: "v1.11.1"
   DOCKER_SERVICE_DOMAIN: "cluster.local"
